Add method for cost/conversion rate
Created method and tests at service level
created Endpoint and test

+++ src/main/java/com/spaceboost/challenge/model/AdGroup.java
@@ -18,5 +18,5 @@ public class AdGroup {
     private int campaignId;
     private int clicks;
     private int conversions;
-    private float cost;
+    private Double cost;
 }
