+++ src/main/java/com/spaceboost/challenge/model/Keyword.java
@@ -21,6 +21,6 @@ public class Keyword {
     private int adGroupId;
     private int clicks;
     private int conversions;
-    private float cost;
+    private Double cost;
 
 }
